Safari-611.1.10/LayoutTests/http/tests/preconnect/link-header-rel-preconnect-http-expected.txt

CONSOLE MESSAGE: Successfuly preconnected to http://localhost:8000/
Tests that Link header's rel=preconnect works as expected over HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
